desktop: ensure desktop directory on application init
Until now we were creating the desktop directory in a lazy way, like
any other cached directory.

However, we have the problem that at some point we have to dispatch
between different types of files, when creating them for the cache.

We cannot know when we will need to create the desktop directory for
first time in order to discern between that type of directory, or the
regular one.

What we can do is ensure that we created the desktop directory before
any other part of nautilus request it. In this way, we can create it
on our subclasses of the desktop, and after that, nautilus will request
the cache as a regular use, without the need to special case the
desktop.

For that, create the desktop directory when the desktop application
starts, holding the reference so the cache doesn't release it, and then
let nautilus work as expected.

For that, in previous commits we moved the file dispatching to be inside
the directory, so now any file creation happens inside the directory,
and therefore we can control, when creating the desktop directory, what
subclass will be called.

https://bugzilla.gnome.org/show_bug.cgi?id=712620