file-dnd: move dnd handling to file
We are requiring file handling like special casing subclasses in the
dnd handling.

We are currently doing it outside of nautilus class... which makes
overriding on subclasses impossible.

This design goes against inheritance design, so there is no point on
making it that way if it defeats this purpose.

So merge the handling of file dnd inside the file class itself, and in
upcoming patches we will override that handling in order to use
inheritance instead of special casing in the parent.

https://bugzilla.gnome.org/show_bug.cgi?id=712620