toolbar: use widget visibility instead of desktop special casing
We need to avoid showing the operations popover in the desktop, or
basically in any window that has the disable-chrome property set.

We can check inside the toolbar if the toolbar itself is hidden in order
to know whether it makes sense to show the operations popover.

In this way we don't relay in special casing subclasses of the window.

https://bugzilla.gnome.org/show_bug.cgi?id=712620