nautilus-dnd: request file system info if not available
We were assuming we had in cache the files used for dnd, but with
the desktop split that's not longer the case.

What can happens is that we cannot discern whether the drop action
should be a copy or a move, since Nautilus always performs a move if
the file system is the same as the dragged files.

To fix it, in a best effort mode, we will request to update the files
information if it's not available. In future times it will be requested,
and if the files are not freed, we will have the file system information
available in order to decide a more appropriate drop action.

https://bugzilla.gnome.org/show_bug.cgi?id=712620