directory: use gio extensions for directory dispatching
We needed to do some hackish code in order to allow types that are
not included in nautilus-directory to dispatch the correct subclass.

Instead of that, we can just create a "plugabble" system that allows
directory types to be registered in the system, and implement a class
vfunc that queries if the class handles a specific type of uri, falling
back if none can handle it to the usual nautilus-directory.

We can do this for the desktop directory and the search directory.

https://bugzilla.gnome.org/show_bug.cgi?id=712620