window-slot: fix condition for disconnecting old view
We were checking for the new view instead of the old view...
Needless to say that's wrong, and we were not disconnecting at all.

It was not crashing because the view is disconnected when is going to be
destroyed, so no more signals are done.

However, when we were destroying the view, the view signals the
end-loading signal, which forces the slot to free any change data, in
middle of a change.

"Luckely" thanks to all the safe checks in window slot, this was not
crashing, but just misbehaving and not updating the window with the new
view data.

https://bugzilla.gnome.org/show_bug.cgi?id=764981