desktop: add command line options for override gsettings
Sometimes we want to override the show-desktop-icons gsettings, as we
were doing before splitting the desktop.

Wrongly I assumed that since it's a different binary, once can simply
run it or not, but of course that was an oversimplification, and forgot
all what I needed to do in order to support all the cases for the
desktop handling.

This patch adds the missing command line options we had, --force-desktop
and --no-desktop, and also adds the --force-desktop to the classic
desktop file, since we needed to enable the classic mode.

https://bugzilla.gnome.org/show_bug.cgi?id=765159