toolbar: convert action menu from GMenu into GtkPopoverMenu
This is in preparation for merging the view menu into the action menu.
The action menu is currently a GMenu, which doesn't support the controls
we have on the view menu, e.g. the zoom slider and the button rows.
Converting to a GtkPopoverMenu will allow us to merge the menus.

This is part of the toolbar menu redesign to improve the usability of
the menus.

https://bugzilla.gnome.org/show_bug.cgi?id=764632