files-view: remove action group only matches view group
We were removing the action group when the parent is unset.
However, this removes the action group from the window unconditionally,
meaning that if the view that the parent is unset for is not the current
view it will remove the action group of the actual current view, making
the actions unreachable outside of the view.

This was causing the action menu in the toolbar to have non activatable
actions.

To fix it, make sure we remove the action group only if it matches the
current one attached to the window.

https://bugzilla.gnome.org/show_bug.cgi?id=765489