files-view: select first item only when searching
Since 2fa597e we select the first item even if the view didn't load
all the files.

But we were selecting the first item unconditionally.

However, it only makes sense to select the first item on the view if the
view is searching, not otherwise.

This patch makes files view to select the first item only if the view is
searching, as we had previously.

https://bugzilla.gnome.org/show_bug.cgi?id=765264