directory, file: add function to query FUSE handling
We use checks to know whether a file is local, using g_file_is_native.
However, we also need to know for some cases when a file, even if no
native, can be handled like native, using FUSE.

This will be needed for the selection of default applications on remote
locations, and will be used in a upcoming patch.

https://bugzilla.gnome.org/show_bug.cgi?id=764506