mime-actions: check for fuse for default applications
We were checking if the file is native, and if not, we request the
default applications that they must handle uris.

However, we can use FUSE in order to convert those uris to local paths
and therefore use a wider variety of applications.

This fix some default application not being available in network
locations.

https://bugzilla.gnome.org/show_bug.cgi?id=764506