files-view: reveal file after renaming or creating
We want to reveal the file that was just renamed.
In order to do that, we have to make sure the view has the icon and all
changes were acknowledge, and only then, reveal the file.

In this patch, we implement this creating a hash table that will hold
any file that needs to be revealed when the view acknowledges it.

Functions using this will be the renaming of files, the creation of
folders, and the creation of folders with selection.

https://bugzilla.gnome.org/show_bug.cgi?id=764137