application: Fix white desktop after changing themes
Changing the theme to Adwaita was causing the Adwaita.css style provider
to get added after the nautilus.css style provider. This caused the
Adwaita.css provider to override the nautilus.css provider, because the
behavior of two style providers with the same priority is undefined and
gtk happens to prefer the style provider that got added last.

https://bugzilla.gnome.org/show_bug.cgi?id=765687