file-undo-operations: use queue for storing sources and destinations
When copying, moving or linking files, original sources and their destinations
are stored in an information structure used for undoing the operation. The
sources and destinations were appended at the end of a list. Due to append
taking linear time, these file operations would have a resulting quadratic
complexity depending on the number of files. When operating on thousands of
files, this would lead to a significant decrease in speed. In order to fix
this, the sources and destinations are stored in a queue that allows appending
in constant time.

https://bugzilla.gnome.org/show_bug.cgi?id=757747