file-operations: Use a GList** for the to_delete output parameter
The to_delete GList* gets initialized in trash_files() and is supposed
to be filled by trash_file() so that we know which files we want to
delete directly when the call to g_file_trash() fails for some reason.

However, this GList* gets modified in place by calls to g_list_prepend()
and so those changes will only be seen by the to_delete argument inside
trash_file(), but not by the local variable defined in trash_files(), which
will continue to point to NULL regardless of those changes.

So, we simply need to pass a reference to the GList* so that any change done
to the structure of the GList* inside trash_file() is seen in trash_files().

https://bugzilla.gnome.org/show_bug.cgi?id=766030