properties-window: make icon chooser dialog modal
The custom icon selection GtkFileChooserDialog is unresponsive when
opened. This is caused by the properties dialog window being modal,
since it is up to the window manager how to process input in such
situations.

One of the ways to ensure that the child dialog receives and processes
user input is by making it modal. This commit makes it such.

https://bugzilla.gnome.org/show_bug.cgi?id=753394