files-view: fix renaming logic
When renaming a file from the popover, the code checks if there is an
existing file with the name entered and the name entered does not match
the name of the file being renamed. That has a side effect of allowing
the original filename to be entered without reporting an error.

Some problems arise when the model includes files from different
directories (e.g. when searching or expanding directories in the list
view).
When searching, the code will not allow a rename if there are files
with the entered name, despite being in separate directories.
When renaming a file in an expanded directory, the code will ignore
files inside it and only check the currently opened one. This also
applies when creating new folders.

This commit fixes that by checking for duplicate names inside the parent
directory when renaming and checking the directory at
nautilus_file_view_get_backing_uri() when creating new folders.

https://bugzilla.gnome.org/show_bug.cgi?id=757385