pathbar: fix underallocations
Starting with GTK+ 3.21, Nautilus is constantly reporting
underallocations happening in the path bar. That is chiefly caused by
some assumptions made in the code (i.e. both slider buttons are of the
same width, which is untrue, due to the “linked” style class).

This commit corrects those assumptions by using widths of both slider
buttons in allocation calculations.

https://bugzilla.gnome.org/show_bug.cgi?id=768916