query-editor: unref file in update_information_label()
window-slot binds with query-editor via a property, which results in
nautilus_query_editor_set_location() being called on construction and
location changes. That function in turn calls
update_information_label(), which takes a ref on the file, but does not
release it. This can result in a crash with remote bookmarks when
connecting/disconnecting or switching back and forth.

This commit fixes that by adding an unref to update_information_label().