files-view: move file name widget logic to separate controllers
The rename file popover and the new folder dialog share common logic for
validating file names entered by the user. The control logic was implemented
with a simple structure in files-view. Besides common logic, the structure also
held parameters specific to only one of the operations. Another problem is that
the current implementation does not allow flexibility in obtaining the file name
from the widgets and displaying error messages.

In order to fix this, reimplement the structure as an abstract class and create
two subclasses for the "Rename" and "New Folder" widgets.

https://bugzilla.gnome.org/show_bug.cgi?id=769336