file-operations: update progress on skip
Currently, the transfer info of an operation is only modified if it is
successful, resulting in a confusing reflection in the UI. Treating
skipped operations as completed and displaying them as such feels more
natural, as they are technically completed (i.e. nothing has been done).
This commit changes the behavior as such.

https://bugzilla.gnome.org/show_bug.cgi?id=769383