file-conflict-dialog: separate file logic from UI management
In Nautilus, file conflicts are handled by a specific dialog. Previously, the
dialog class managed both the UI and the related nautilus files. This lead to it
being inflexible due to operation specific logic being mixed with the rest of
the functionality. In order to change this, move file logic to a separate module
and add methods for controlling the UI elements in the dialog.

Create an operation-manager module to handle dialog controlling. Move anything
related to Nautilus files from the dialog class to the new module.

https://bugzilla.gnome.org/show_bug.cgi?id=770160