Implement batch renaming
Renaming multiple files at once has been a missing feature in Nautilus
for a long time. This patch implements that feature in the following way:

This operation is launched in the same way as the rename one, when the
selection has more than one file.

When the batch renaming is launched, a dialog is shown, offering two
modes.

In the first mode, the user can use metadata (if available), numbering and
original file name tags to create the new names. Between the tags, there
also can be written normal text, which will be added in the new names. If
numbering is used, the order of the files can be modified by using several
criteria.

In the second mode, the user can replace an existing part of the name.

https://bugzilla.gnome.org/show_bug.cgi?id=768311