shell-search-provider: remove simple engine recursivity
In 26a3ab4cd5 we put recursivity to the simple search engine in order
to fix the report of some users that weren't getting files inside
folder of home directory.

However this is wrong, since it takes too long and gnome-shell search
provider doesn't have a way to get partial results or close the search
when done, which makes Nautilus not reporting results until the search
is done, and in case the user stops the shell search, nautilus keeps
searching making the CPU consumption really high.

In fact, we always made recursivity for tracker but not for the simple
engine when performing the shell search.

The issue users were experiencing seems unrelated to this, and I
actually cannot reproduce, so let's return to the regular behaviour
we had.

https://bugzilla.gnome.org/show_bug.cgi?id=766174