mime-actions: Fix shift+control+down seg fault on a folder
Using this shortcut on one or more folders causes segmentation fault.

In order to solve this, if there is a directory that needs to be
activated, the NEW_WINDOW flag is set, so that the initial window will
be closed and a new one will be created. After a new window is created,
the CLOSE_BEHIND flag is disabled,so that the window will not be closed.
The flags NEW_TAB is also set there(unless we want all directories
opened in new windows), in order for new tabs to open in the newly
created window.

https://bugzilla.gnome.org/show_bug.cgi?id=755711