batch-rename-dialog: fix search & replace for . and ..
In the replace mode, when "." or ".." were typed in the replace entry,
there would always pop up an error, saying that a file would have an
invalid name.

In this case, instead of checking the entry, the check must actually be
done to all the new names to see if any of those are equal to "." or
"..".

Since the new names are now used in this function, these have to be
obtained before.

https://bugzilla.gnome.org/show_bug.cgi?id=770870