batch-rename-dialog: use GTask appropiately
The Gtask has to finish only when all the internal async call are ready,
doing the other way around is conceptually wrong and will cause problems
if we rely on the task for checking the status of the general operation.

For that implement a correct GTask handling, waitig with a mutex and
condition for all async calls to be ready, and only then mark the task
as finished.

https://bugzilla.gnome.org/show_bug.cgi?id=770586