batch-rename-utilities: fix unicode chars for replace
In the replace mode, for some unicode characters, the displayed label
would be incorrect.

The problem is that the second argument of g_markup_escape_text is
the number of bytes and it was used with the number of characters.

Since the strings are null terminated we can just omit this parameter
altogether and fix this as a side effect.

https://bugzilla.gnome.org/show_bug.cgi?id=770972