nautilus-file: Fix segfault on batch rename for file gone
If one or more files are deleted during a batch rename and click rename,
Nautilus would seg fault.

The problem is that in the function nautilus_file_rename_handle_file_gone
there was called a callback that was NULL. For batch renaming, there is no
need for a callback here.

https://bugzilla.gnome.org/show_bug.cgi?id=770968