nautilus-file: Fix segfault on undo for batch rename
In some cases, if some of the files are not renamed, for reasons like
deletion or files having the same name, the undo would fail.

The problem was that the files and names lists provided for undo also
contained info about the files for which the rename failed.

The fix was to add undo info only for files that are actually renamed
and prepare the undo only if at least a file has successfully been
renamed.

https://bugzilla.gnome.org/show_bug.cgi?id=770965