batch-renaming: disable if compiled without tracker
We have a compile option to not use tracker. Batch renaming relies on
tracker for part of its functionality, and disabling that would mess the
code with ifdefs.

I think doesn't make much sense to have a compile time switch for
tracker anymore (tracker can still be disabled in runtime and it's fine)
so I would like to remove that switch.

Since it's late in the cycle, better to not do that, but let's try for
next release. For now just make the batch-renaming disabled in case
the compilation was done without tracker.

https://bugzilla.gnome.org/show_bug.cgi?id=771158