batch-rename-dialog: don't multi-thread NautilusDirectory calls
Even if late... I just realized neither NautilusDirectory or
NautilusFile are thread safe.

We were using them from two different threads for checking conflicts,
and that was crashing when the internal hash table was modified and
accessed from the different threads.

The ideal fix would be to make NautilusDirectory and NautilusFile thread
safe, but that's a big work. So for now, call the main thread for using
them.

https://bugzilla.gnome.org/show_bug.cgi?id=771129