nautilus-file: copy parameter data in batch renaming
We need to have our own copy and ownership of the parameter data,
if not later on when we do async operations the data might be already
freed.

This was making batch rename crash every time.

This is to follow the common pattern of not transferring the ownership
from caller to called.

https://bugzilla.gnome.org/show_bug.cgi?id=770586