batch-rename-dialog: use strlen for gtk_editable_insert_text
We were using utf8_lenght, which is the right approach to know the
number of characters. However, gtk_editable_insert_text expects bytes
rather than number of characters.

This was causing languages that have unicode characters in the tags
to have a buggy batch rename dialog.

To fix it, use strlen so we get the number of bytes.

https://bugzilla.gnome.org/show_bug.cgi?id=772063