batch-rename-dialog: revealer for auto numbering
Currently we have a big padding that initially has no meaning.

The problem is that when using the rename for multiple files, we initially
have an extra padding due to auto numbering. Even if the automatic numbering
is not used, the space for the toggle button used for the numbering options
is now blank and waiting for a possible use of the automatic option.

Add a GtkReaveler that contains a child represented by a GtkBox. The GtkBox
contains a label and the toggle button for choosing an option of automatic
renaming. The child is revealed only when necesarry, more specifically when
the automatic numbering option is added.

https://bugzilla.gnome.org/show_bug.cgi?id=771813