batch rename: show name when hovering over files
When we rename a group of files, we can see the names of the files
inside a window with a certain dimension, so we see the part of the name
that fits the window.

The problem is that we cannot see the entire name of a file if the name
is too long.

Add a tooltip text on the label containing the name of a file.

https://bugzilla.gnome.org/show_bug.cgi?id=771467