file-operations: avoid scanning the same files
In some situations when operating on a set of files, the same file can be
scanned more than once. An example would be when the input set of files contains
a directory's child before the directory itself. This leads to wrong progress
information being displayed. In order to fix this, keep a set of the files
scanned so far.

https://bugzilla.gnome.org/show_bug.cgi?id=766027