file: add function nautilus_file_is_regular_file
In commit we added a new function named nautilus_file_is_regular_file

The problem is that there are function present in different files
named "!nautilus_file_is_directory()" like inside nautilus-file.c
near line #7798. So, if any newcomer will come and look into the files
then it is difficult for them to analyse that there is any other
option available for file or not.

To solve this issue create a new function inside nautilus-file
with name nautilus_file_is_a_regular_file. So, we can check
for explicit regular file type.

https://bugzilla.gnome.org/show_bug.cgi?id=772316