list-view: don't crash on view leaks
Although we all love how Nautilus keeps us on the belt crashing when
there is a leak, users are not happy about that.

So let's make nautilus more docile and don't crash in case there is a
view leak. For that, disconnect all signals in the dispose handler
rather than in finalize.

https://bugzilla.gnome.org/show_bug.cgi?id=772462