ui-utilities: use gtk_menu_popup_at_pointer instead of gtk_popup_menu
We were using gtk_popup_menu which is problematic in backends with no
global coordinates like wayland.

To fix it, we can use gtk_menu_popup_at_pointer which uses the event
relative coordinates.

Also, seems the custom handling we were doing for the popup menu in that
code is obsoleted, so there is not need to specify the button.

https://bugzilla.gnome.org/show_bug.cgi?id=772507