batch-rename : Reword "Rename %d files/folders" as "Rename %d files"
The problem is that when we select both folder and file then string
"Rename %d files" appears which may not be the right term when
the selected items contains both files and folders.

To solve this issue we rename the string and now in place of string
"Rename %d files" we use "Rename %d Files and Folders" for both file
and folder.

https://bugzilla.gnome.org/show_bug.cgi?id=771586