files-view: Use <control>Return to open selected folder in new window
It makes more sense than the complex <control><shift>w accel.
This is the default behavior for launching a new instance of an
opened application in GNOME Shell.
Currently, pressing "Return/Enter" with a selected folder, opens
it. In doing so, using the <control> modifier should open the
selected folder in a new window.
This widespread behavior is also related to <control> + click for
opening links in a web browser.

https://bugzilla.gnome.org/show_bug.cgi?id=772925