floating-bar: fix hide on hover with notebook visible
When the notebook is visible, the pointer intersection with the floating
bar check fails. That is due to how the floating bar bounds are
calculated - the offset of the overlay when tabs are visible is not
taken into account, allowing the upper y bound to become smaller than
the lower bound. That can be fixed by calculating the upper y bound
relative to the lower one.

https://bugzilla.gnome.org/show_bug.cgi?id=773119