undo-file-manager: make the class final and port to G_DECLARE* type declaration
Changed the old GObject class declarations in order to use smart pointers to
generate some functionalities and make the code more readable.

The class should not be derivable as its` functionality logics requires it to be
final.

To fix this use G_DECLARE* type and make the class final instead of derivable.

https://bugzilla.gnome.org/show_bug.cgi?id=773831