mime-actions: support admin backend
Until now Nautilus was not able to handle files where the user had
no permissions. An error was reported.
The only way for a user to handle those files were to start Nautilus
with sudo, which is something that shoudl be avoided for security
reasons.

On Wayland, is not even possible to launch an application with sudo,
so this is no longer available and therefor no way to handle files
with no permissions.

On 3.22 gvfs added an admin backend with integration with Polkit,
so a file withouth permissions can be accessed using this backend
if the user has the root password.

Add support for the admin backend in Nautilus, where a file will
be opened using it if cannot be read.

There still work to do, basically implement the operations with
this backend too and refactor the code to be able to open from
nautilus application command line also a file withouth permissions.

https://bugzilla.gnome.org/show_bug.cgi?id=773937