view: make icon getter static
We were relying on the current view to return the icon to put in the
toolbar, however that requires a view instance and also we cannot
control really what icon we want to show in which circumstances.

We want more control about that, so we need a single entry point where
we can get the icon to show depending on the known view types we
have.

So this patch converts the view property to a static method.

https://bugzilla.gnome.org/show_bug.cgi?id=771075