window-slot: reverse the view icon action
That means, now we will show the grid icon when the current view is
the list view and the opposite.

This goes in similarity with gnome-boxes.

The main question to decide this is, is the button an action, and
therefore the icon should be the "target" of the action, or is more
about an information of the current state, and therefore the icon should
be about the current view type?

This patch decides the icon represents an action, although would be good
to have some user testing.

https://bugzilla.gnome.org/show_bug.cgi?id=771075