progress-info: avoid races
We were not checking whether the operations was aborted from the user
side when setting the details,so that could make a race and the info
thread setting the details to canceled and the operation thread setting
the details too afterwards before realizing the operations was canceled.

To fix it, always check whether the info progress was canceled or not.

https://bugzilla.gnome.org/show_bug.cgi?id=775094