file-operations: Fix progress of copies over gvfs
commit 24dae9e0a "file-operations: avoid scanning the same files"
added a way to skip files which have already been processed
in scan_dir(). To do that, the code uses G_FILE_ATTRIBUTE_ID_FILE,
which does not seem to be widely implemented by gvfs modules.

When the file has no ID, it will not be accounted for in the SourceInfo
structure used to track copy progress, resulting in progress not being
shown, and Nautilus saying that it's copying 0 files.

This commit changes the logic to use the file URI instead of the file ID
to check if the  file has already been processed.