file-operations: don't recurse for trash operations
We were scanning the sources as we do for copy and move operations,
and to count the files we were recursing inside directories.

However, the trash operation doesn't recurse at all, so the operation
progress was wrong.

Instead of that, don't recurse for counting the number of files, so the
trash progress is correct. Although is not coherent with the rest of the
operations, recursing for just showing all the files in the progress
would be extremely counterproductive.

https://bugzilla.gnome.org/show_bug.cgi?id=775094