files-view: remove an unnecessary helper function
selection_not_empty_in_menu_callback was simply a verbose way to test
whether the selection GList was NULL, which is tidier to just do inline.

The call giving this function its name would be out-of-order if it were
really needed, but it's not since can_set_wallpaper already checks this.