window: remove max forced window size calculation
Currently, the default size for window is calculated by taking the
minimum value between the screen size and the hardcoded default. As
the smallest recommended screen size for GNOME today is 1024x600 and the
code has been added in 2000, it is likely no longer relevant. This
commit simply makes the code use the hardcoded values within the
recommended range as the default window size.

https://bugzilla.gnome.org/show_bug.cgi?id=779082