In this commit, we added CTRL+m as a shortcut for renaming
In order to rename a file, F2 was used as a shortcut. However, some keyboards might not have F's and an extra shortcut was thus needed(one that would preferably be CTRL+m).
In order to fix this, I added a "const gchar *rename_fil[]" consisting of F2 and <ctrl>m so that both of these are available for renaming. This change should work in pretty much any case.

https://bugzilla.gnome.org/show_bug.cgi?id=763727