general: add recency sort for recent files
The sort order is based on atime currently, which is problematic,
because some daemons (i.e. dropbox) randomly accesses files and
changes atime...
Instead we should just take into account when the user accessed.

Recently glib and gvfs added a new attribute in the recent namespaces
G_FILE_ATTRIBUTE_RECENT_MODIFIED or "recent::modified" that we can use
to sort correctly the files in Recent.

This patch adds this attribute and corresponding columns etc. to the
file data and views and makes it the default sorting when in Recent.

https://bugzilla.gnome.org/show_bug.cgi?id=777507