batch-rename: fix stack child height
With numbering added, the format mode stack child is slightly smaller
than the replace one, so the scrolled window height is slightly
changing when switching between modes.

The problem is that the row spacing in the replace stack child is
greater by 1px than the spacing in the format stack child.

To fix this make the replace stack child have the same row spacing.

https://bugzilla.gnome.org/show_bug.cgi?id=778272