window: Don't use eel to show a dialog
Most of the things in eel are already in glib, or are already easy
enough to have them directly in the code. So we should remove eel which
is just another layer of abstraction that we don't need as it makes
it hard to follow the code.

This patch replaces eel_show_error_dialog with show_error_dialog
to make it independent of eel.

https://bugzilla.gnome.org/show_bug.cgi?id=775092