toolbar: add theatrics animation to the operations button
We have been going back and forth of a solution to make the operations
button noticeable enough once it appears.

First we implemented a highlight inside the button. But that wasn't
enough, so we had to show the operations popover at the start.

That is not really good because it appears in every window, and the user
has to explicitly close it, which is annoying.

This patch implements another animation highlight, which is more
visually bold to try to catch the attention, in order to not need to
show the operations popover.

https://bugzilla.gnome.org/show_bug.cgi?id=753728