file-operations: use standard print functions
We were using a print function with custom handlers.

Not only this function is hard to understand and follow, but is also non
standard which makes it problematic for translation (where everything is
check using c standards to avoid crashing the application if the translator
makes some unsuported change).

This patch removes every use case of f() function and uses g_strdup_printf
and g_strdup in its place and uses custom functions directly to convert
required parameters into gchar* types.

https://bugzilla.gnome.org/show_bug.cgi?id=775091