preferences-window: Use spin button for thumbnail limit
File sizes (especially compressed images) are arbitrary values that are
highly dependant on the camera resolution, and contents of the image.
Therefore it does not make sense to have this be a drop-down with preset
hard-coded values: A 21MP camera often produces 12MB images... but the
only option is 10MB, or 100MB.

This commit changes the combobox in the preferences to a spinbutton and
changes the signal handler for thumbnail-limit.
The value is in MB, with a maximum of 4096 and a minimum of 1.

https://bugzilla.gnome.org/show_bug.cgi?id=779165